backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
pymysql
sqlalchemy
requests
ddgs
dbutils
passlib[bcrypt]
pydantic[email]